Adds basic style for page
