+++ src/graphics/deprecated/main.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Main Layout</title>
+    <link href="../css/theme.css" rel="stylesheet" type="text/css" />
+  </head>
+  <body>
+    <div id="container"></div>
+    <script src="./main.tsx" type="module"></script>
+  </body>
+</html>
