+++ src/graphics/main.html
@@ -6,7 +6,7 @@
     <link href="css/theme.css" rel="stylesheet" type="text/css" />
   </head>
   <body>
-    <div id="container"></div>
-    <script src="./main.tsx" type="module"></script>
+    <div id="app"></div>
+    <script src="main.tsx" type="module"></script>
   </body>
 </html>
